Rename editor to EditorContainer
